+++ services/backend/src/routes/index.ts
@@ -9,10 +9,16 @@ import { IRoute } from "../lib/IRoute";
 import { routes as exampleRoutes } from "./example";
 import { routes as user } from "./user";
 import { routes as sessionRoutes } from "./session";
+import { routes as meRoutes } from "./me";
 
 // =======================================================================
 
-const routes: IRoute[] = [...exampleRoutes, ...user, ...sessionRoutes];
+const routes: IRoute[] = [
+  ...exampleRoutes,
+  ...user,
+  ...sessionRoutes,
+  ...meRoutes,
+];
 const router = Router();
 routes.forEach((route) => {
   route.buildRoute(router);
